+++ src/content/index.ts
@@ -1,7 +1,9 @@
 import { mountDropboxExplorer } from "./dropboxExplorer";
+import { mountGitHubExplorer } from "./githubExplorer";
 
 function mount(): void {
   mountDropboxExplorer();
+  mountGitHubExplorer();
   void browser.runtime.sendMessage({
     type: "primitive-io:content-mounted",
     href: location.href